pyproject.toml
[tool.poetry]
name = "meri"
version = "0.1.0"
description = "Meri 🌊 – Klikkikuri backend service for news title unsensationalizer"
authors = [
	"Teemu Autto <teemu.a.autto@jyu.fi>",
	"Juho Tarkkanen <juho.tarkkanen@hotmail.fi>",
	"Taija Kolehmainen <taija.s.kolehmainen@jyu.fi>"
]
packages = [
	{ include = "valas", from = "src/" },
	{ include = "meri", from = "src/" }
]
readme = "README.md"
repository = "https://github.com/Klikkikuri/meri"

[tool.poetry_bumpversion.file."src/meri/__init__.py"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
ignore = ["E701"]

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
openai = "^1.51.1"
Jinja2 = "^3.1.4"
structlog = "^24.4.0"
opentelemetry-api = "^1.27.0"
playwright = "^1.47.0"
opentelemetry-sdk = "^1.27.0"
newspaper4k = "^0.9.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.9"
poetry = "1.8.3"
isort = "^5.13.2"
pre-commit = "^4.0.1"
bandit = "^1.7.10"
poetry-bumpversion = "^0.3.2"
requests-cache = "^1.2.1"

[tool.poetry.group.otel]
# OpenTelemetry dependencies
optional = true

[tool.poetry.group.otel.dependencies]
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-asyncio = "^0.48b0"
opentelemetry-instrumentation-logging = "^0.48b0"
opentelemetry-instrumentation-requests = "^0.48b0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
